+++ celescope/flv_CR/__init__.py
@@ -1,5 +1,5 @@
 __STEPS__ = ['sample', 'barcode', 'convert', 'assemble', 'annotation', 'match', 'summarize', 'mapping'] 
-__ASSAY__ = 'fl_vdj_CR'
+__ASSAY__ = 'flv_CR'