+++ celescope/flv_CR/mapping.py
@@ -2,7 +2,7 @@ import pandas as pd
 import glob
 from celescope.tools import utils
 from celescope.tools.step import s_common
-from celescope.fl_vdj_CR.VDJ_Mixin import VDJ_Mixin, get_opts_VDJ_Mixin
+from celescope.flv_CR.VDJ_Mixin import VDJ_Mixin, get_opts_VDJ_Mixin
 
 
 class Mapping(VDJ_Mixin):