+++ celescope/flv_CR/match.py
@@ -1,10 +1,10 @@
 import pandas as pd
 import pysam
 
-from celescope.fl_vdj_CR.annotation import Annotation
+from celescope.flv_CR.annotation import Annotation
 from celescope.tools import utils
 from celescope.tools.step import s_common
-from celescope.fl_vdj_CR.VDJ_Mixin import VDJ_Mixin, get_opts_VDJ_Mixin
+from celescope.flv_CR.VDJ_Mixin import VDJ_Mixin, get_opts_VDJ_Mixin
 
 
 class IncorrectMatchDir(Exception):