+++ celescope/flv_CR/multi_flv_CR.py
@@ -23,10 +23,7 @@ class Multi_flv_CR(Multi):
     conda activate celescope
     multi_flv_CR \\
         --mapfile ./test.mapfile \\
-        --chemistry flv \\
-        --mem 10 \\
         --thread 8 \\
-        --allowNoLinker \\
         --seqtype TCR \\
         --ref_path "/SGRNJ/Database/script/soft/cellranger/vdj_ref/6.0.0/hs/refdata-cellranger-vdj-GRCh38-alts-ensembl-5.0.0" \\
         --soft_path "/SGRNJ03/randd/cjj/soft/cellranger/cellranger-6.1.2/cellranger"