+++ celescope/flv_trust4/multi_flv_trust4.py
@@ -10,10 +10,7 @@ class Multi_flv_trust4(Multi):
     ```
     multi_flv_trust4 \\
         --mapfile ./test.mapfile \\
-        --outdir ./ \\
-        --chemistry flv \\
-        --allowNoLinker \\
-        --ref GRCm38 \\
+        --ref hg38 \\
         --thread 8 \\
         --seqtype BCR \\
     ```