+++ docs/CHANGELOG.md
@@ -2,6 +2,7 @@
 ## [1.11.0] - 
  ### `flv_vdj`
  - Chemistry `flv_rna`(full length VDJ matched mRNA) and `flv` can be auto-detected. 
+ - Remove assay `flv_trust4_split`.
 
  ### `snp`,`capture_virus` and `fusion`
  - Change the default `--umi_threshold_method` from `auto` to `otsu`.